+++ includes/referral-fee-calculator.php
@@ -5,7 +5,7 @@ function calculate_referral_fees($author_id = null, $inquiry_status = 'pending')
     $referral_fee_type = get_option('pdr_referral_fee_type', 'both'); // Default to 'both' if not set
 
     $view_referral_fee = 0.00;
-    $approval_referral_fee = 0.00;
+    $agreement_reached_referral_fee = 0.00;
 
     // Determinar se deve usar as configurações do autor ou as globais
     $use_author_settings = $author_id && get_user_meta($author_id, 'pdr_override_referral_fee', true) === 'yes';
@@ -17,23 +17,24 @@ function calculate_referral_fees($author_id = null, $inquiry_status = 'pending')
     if ($referral_fee_type === 'view' || $referral_fee_type === 'both') {
         $view_referral_fee = $use_author_settings ? get_user_meta($author_id, 'pdr_referral_fee_view', true) : get_option('pdr_general_referral_fee_view', '0.05');
     } else {
-        $view_referral_fee = 0.00; // Não calcular Referral Fee por visualização se o tipo for 'approval' apenas
+        $view_referral_fee = 0.00; // Não calcular Referral Fee por visualização se o tipo for 'agreement_reached' apenas
     }
 
-    if (($referral_fee_type === 'approval' || $referral_fee_type === 'both') && $inquiry_status === 'approved') {
-        $approval_referral_fee = $use_author_settings ? get_user_meta($author_id, 'pdr_referral_fee_approval', true) : get_option('pdr_general_referral_fee_approval', '0.10');
+    if (($referral_fee_type === 'agreement_reached' || $referral_fee_type === 'both') && $inquiry_status === 'agreement_reached') {
+        $agreement_reached_referral_fee = $use_author_settings ? get_user_meta($author_id, 'pdr_referral_fee_agreement_reached', true) : get_option('pdr_general_referral_fee_agreement_reached', '0.10');
     } else {
-        $approval_referral_fee = 0.00; // Não calcular Referral Fee por aprovação se não está aprovado ou se o tipo é 'view' apenas
+        $agreement_reached_referral_fee = 0.00; // Não calcular Referral Fee por acordo se não está aprovado ou se o tipo é 'view' apenas
     }
 
     // Normalizar e formatar os valores decimais
     $view_referral_fee = number_format((float)str_replace(',', '.', $view_referral_fee), 2, '.', '');
-    $approval_referral_fee = number_format((float)str_replace(',', '.', $approval_referral_fee), 2, '.', '');
+    $agreement_reached_referral_fee = number_format((float)str_replace(',', '.', $agreement_reached_referral_fee), 2, '.', '');
 
     $referralFees = [
         'view' => $view_referral_fee,
-        'approval' => $approval_referral_fee
+        'agreement_reached' => $agreement_reached_referral_fee
     ];
 
     return $referralFees;
 }
+?>